@samj1912 2024-09-20
> Does this require all servers to be upgraded to a new release of Kine before the schema can be changed, or is this change safe to make without updating all the nodes?

It is safe to make the change without updating all the nodes. The queries are b/w compatible since we didn't create any new columns.