@samj1912 2024-09-20
Looks like cockroachdb does not support per column collations?