@samj1912 2024-09-20
Actually on a closer look it looks like it supports column collation, it just does not support `C` as a collation?