@brandond 2024-09-20
Yeah, it looks like it's using `golang.org/x/text/language` and ends up calling something like `v, err := language.Parse("C")`?