@samj1912 2024-09-20
I might end up doing something along the lines of 

```sql
select version();
```

and changing the column type based on whether that contains cockroach vs postgres. Is that reasonable? For unknown version strings, I will just assume postgres.