@samj1912 2024-09-20
Some comparisons from PR perf tests - 


### List Before

> [postgres-12.16] [PERF]	0.034  average postgres-12.16 request duration (seconds): list configmaps
[postgres-12.16] [PERF]	0.005  [  32]  █▌
[postgres-12.16] [PERF]	0.025  [ 296]  ██████████████▏
[postgres-12.16] [PERF]	0.05   [1045]  ██████████████████████████████████████████████████
[postgres-12.16] [PERF]	0.1    [ 212]  ██████████▏


### List After

> [postgres-12.16] [PERF]	0.027  average postgres-12.16 request duration (seconds): list configmaps
[postgres-12.16] [PERF]	0.005  [ 35]  █▉
[postgres-12.16] [PERF]	0.025  [560]  █████████████████████████████
[postgres-12.16] [PERF]	0.05   [964]  ██████████████████████████████████████████████████
[postgres-12.16] [PERF]	0.1    [ 16]  ▉