@samj1912 2024-09-21
Hmm, not sure why this is still happening, let me check

> building kine: ERROR: at or near \",\": syntax error: invalid locale C: language: tag is not well-formed (SQLSTATE 42601)