@samj1912 2024-09-21
Ok, looks like we are all good 🎉 I will deal with the cockroachdb issue creation later.