@samj1912 2024-09-23
Thanks @brandond for the merge. Can you also cut a release please? Should this be 0.13.0 given the schema migrations needed?