@brandond 2024-09-24
I was waiting to merge another PR, but yes we can tag a 0.13.